requirements.txt
# Web Framework & Server
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Database & ORM
sqlmodel==0.0.16
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation & Settings
pydantic[email]==2.6.1
pydantic-settings==2.2.1
email-validator==2.1.0

# Security & Utilities
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Firebase
firebase-admin==6.4.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.27.0